refactor:winningLottoTicket 도메인 객체 생성
