+++ src/main/kotlin/lotto/LottoController.kt
@@ -2,8 +2,9 @@ package lotto
 
 import lotto.domain.Lotto
 import lotto.domain.LottoNumber
-import lotto.domain.service.LottoGenerator
-import lotto.domain.service.LottoRankFinder
+import lotto.domain.WinningLottoTicket
+import lotto.service.LottoGenerator
+import lotto.service.LottoRankFinder
 import lotto.view.LottoView
 
 class LottoController(
@@ -19,9 +20,10 @@ class LottoController(
 
         val userInputWinningLotto = lottoView.getWinningLotto()
         val winningLotto = Lotto(userInputWinningLotto.map { LottoNumber.of(it) })
-        val bonusNum = lottoView.getBonusNum()
+        val bonusNum = LottoNumber.of(lottoView.getBonusNum())
+        val winningLottoTicket = WinningLottoTicket(winningLotto, bonusNum)
 
-        val rankMap = lottoRankFinder.findLottoRanks(manyLotto, winningLotto, bonusNum)
+        val rankMap = lottoRankFinder.findLottoRanks(manyLotto, winningLottoTicket)
         lottoView.printResult(rankMap)
     }
 }
